// Qanat.Web/src/app/shared/generated/model/water-account-fallow-status-dto.ts
export class WaterAccountFallowStatusDto { 
    WaterAccountFallowStatusID?: number | null;
    Geography?: GeographyMinimalDto;
    SelfReportStatus?: SelfReportStatusSimpleDto;
    ReportingPeriod?: ReportingPeriodSimpleDto;
    WaterAccount?: WaterAccountMinimalDto;
    UsageLocations?: Array<UsageLocationDto> | null;
    readonly CountOfFallowedUsageLocations?: number;
    readonly AcresFallowed?: number;
    CurrentUserCanEdit?: boolean;
    SubmittedByUser?: UserWithFullNameDto;
    SubmittedDate?: string | null;
    ApprovedByUser?: UserWithFullNameDto;
    ApprovedDate?: string | null;
    ReturnedByUser?: UserWithFullNameDto;
    ReturnedDate?: string | null;
    CreateUser?: UserWithFullNameDto;
    CreateDate?: string | null;
    UpdateUser?: UserWithFullNameDto;
    UpdateDate?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}