// Qanat.Web/src/app/shared/generated/model/water-measurement-self-report-simple-dto.ts
export class WaterMeasurementSelfReportSimpleDto { 
    WaterMeasurementSelfReportID?: number;
    GeographyID?: number;
    WaterAccountID?: number;
    WaterMeasurementTypeID?: number;
    ReportingPeriodID?: number;
    WaterMeasurementSelfReportStatusID?: number;
    SubmittedDate?: string | null;
    ApprovedDate?: string | null;
    ReturnedDate?: string | null;
    CreateDate?: string;
    CreateUserID?: number;
    UpdateDate?: string | null;
    UpdateUserID?: number | null;
    WaterAccountNumberAndName?: string | null;
    ReportingPeriodName?: string | null;
    WaterMeasurementTypeName?: string | null;
    WaterMeasurementSelfReportStatusDisplayName?: string | null;
    TotalVolume?: number | null;
    FileCount?: number;
    CreateUserFullName?: string | null;
    UpdateUserFullName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}